Refactor ( use html templates, ng annotate, and john papa's guidelines )
