+++ package.json
@@ -26,10 +26,14 @@
   "homepage": "https://github.com/alenaksu/mdPickers",
   "devDependencies": {
     "gulp": "^3.9.0",
+    "gulp-angular-filesort": "^1.1.1",
     "gulp-autoprefixer": "^3.1.0",
     "gulp-concat": "^2.6.0",
-    "gulp-less": "^3.0.5",
     "gulp-cssnano": "^2.1.0",
+    "gulp-less": "^3.0.5",
+    "gulp-minify-html": "^1.0.6",
+    "gulp-ng-annotate": "^2.0.0",
+    "gulp-ng-html2js": "^0.2.2",
     "gulp-rename": "^1.2.2",
     "gulp-sourcemaps": "^1.6.0",
     "gulp-uglify": "^1.5.1",
