+++ src/components/mdpDatePicker/services/mdp-date-picker.service.js
@@ -1,0 +1,54 @@
+(function() {
+    'use strict';
+
+    angular
+        .module('mdPickers')
+        .factory('mdpDatePickerService', mdpDatePickerService);
+
+    /** @ngInject */
+    function mdpDatePickerService() {
+
+        var service = {
+            formatValidator: formatValidator,
+            minDateValidator: minDateValidator,
+            maxDateValidator: maxDateValidator,
+            filterValidator: filterValidator
+        };
+
+        return service;
+
+        function formatValidator(value, format) {
+            return !value || angular.isDate(value) || moment(value, format, true).isValid();
+        }
+
+        function minDateValidator(value, format, minDate) {
+            minDate = moment(minDate, "YYYY-MM-DD", true);
+            var date = angular.isDate(value) ? moment(value) : moment(value, format, true);
+
+            return !value ||
+                angular.isDate(value) ||
+                !minDate.isValid() ||
+                date.isAfter(minDate);
+        }
+
+        function maxDateValidator(value, format, maxDate) {
+            maxDate = moment(maxDate, "YYYY-MM-DD", true);
+            var date = angular.isDate(value) ? moment(value) : moment(value, format, true);
+
+            return !value ||
+                angular.isDate(value) ||
+                !maxDate.isValid() ||
+                date.isBefore(maxDate);
+        }
+
+        function filterValidator(value, format, filter) {
+            var date = angular.isDate(value) ? moment(value) : moment(value, format, true);
+
+            return !value ||
+                angular.isDate(value) ||
+                !angular.isFunction(filter) ||
+                !filter(date);
+        }
+
+    }
+})();
