+++ src/components/mdpDatePicker/templates/mdp-date-picker-calendar.html
@@ -1,0 +1,26 @@
+<div class="mdp-calendar">
+    <div layout="row" layout-align="space-between center">
+        <md-button aria-label="previous month" class="md-icon-button" ng-click="calendar.prevMonth()">
+            <md-icon md-svg-icon="mdp-chevron-left"></md-icon>
+        </md-button>
+        <div class="mdp-calendar-monthyear" ng-show="!calendar.animating">{{ calendar.date.format("MMMM YYYY") }}</div>
+        <md-button aria-label="next month" class="md-icon-button" ng-click="calendar.nextMonth()">
+            <md-icon md-svg-icon="mdp-chevron-right"></md-icon>
+        </md-button>
+    </div>
+    <div layout="row" layout-align="space-around center" class="mdp-calendar-week-days" ng-show="!calendar.animating">
+        <div layout layout-align="center center" ng-repeat="d in ::calendar.weekDays track by $index">{{ ::d }}</div>
+    </div>
+    <div layout="row" layout-align="start center" layout-wrap class="mdp-calendar-days" ng-class="{ 'mdp-animate-next': calendar.animating }" ng-show="!calendar.animating" md-swipe-left="calendar.nextMonth()" md-swipe-right="calendar.prevMonth()">
+        <div layout layout-align="center center" ng-repeat-start="day in calendar.daysInMonth track by $index" ng-class="{ 'mdp-day-placeholder': !day }">
+            <md-button
+                class="md-icon-button md-raised"
+                aria-label="Select day"
+                ng-if="day"
+                ng-class="{ 'md-accent': calendar.date.date() == day.value }"
+                ng-click="calendar.selectDate(day.value)"
+                ng-disabled="!day.enabled">{{ ::day.value }}</md-button>
+        </div>
+        <div flex="100" ng-if="($index + 1) % 7 == 0" ng-repeat-end></div>
+    </div>
+</div>
