+++ src/components/mdpTimePicker/mdpTimePicker.less
@@ -6,36 +6,36 @@
     height: @clock-size;
     border-radius: 50%;
     cursor: pointer;
-    
+
     padding: (@button-size / 2) + 8;
     background: #ededed;
-	
+
 	.md-button {
 		box-shadow: none !important;
     	background-color: transparent;
 		display: block;
 		position: absolute;
-		
+
         min-height: @button-size;
 		width: @button-size;
         height: @button-size;
-		
+
 		font-size: 12px;
 		line-height: @button-size;
         margin: 0;
         padding: 0;
-        
+
         transform: translate(-50%, -50%);
 	}
-    
+
     .mdp-clock-container {
         width: 100%;
         height: 100%;
         position: relative;
-        
+
         display: block;
     }
-    
+
     .mdp-pointer {
         min-height: 0px;
         width: 1px;
@@ -45,12 +45,12 @@
         right: 0;
         bottom: 0;
         margin: 0 auto;
-        
+
         transform-origin: top center;
         z-index: 0;
         pointer-events: none;
     }
-    
+
     .mdp-clock-center {
         min-height: 0px;
         height: 6px;
@@ -63,7 +63,7 @@
         margin: auto;
         border-radius: 50%;
     }
-    
+
     .md-button.mdp-clock-selected {
         position: absolute;
         border-radius: 50%;
@@ -73,7 +73,7 @@
         left: 0px;
         min-width: 0;
         min-height: 0;
-        
+
         pointer-events: none;
     }
 }
@@ -83,21 +83,21 @@
         padding: 20px;
         width: 309px;
     }
-    
+
     .mdp-timepicker-time {
         padding: 15px;
     }
-    
+
     .mdp-timepicker-selected-time {
         font-size: 3rem;
     }
-    
+
     &:not(.portrait) {
         .mdp-timepicker-time {
             width: 138px;
         }
     }
-    
+
     &.portrait {
         .mdp-timepicker-selected-time {
             font-size: 4rem;
@@ -110,10 +110,7 @@ mdp-time-picker {
     md-input-container.md-has-icon {
         padding-left: 40px;
     }
-    
-    .mdp-time-picker-container {
-    }
-    
+
     .md-button.md-icon-button {
         margin: 18px 0 0 0;
     }
@@ -122,14 +119,14 @@ mdp-time-picker {
 .mdp-timepicker-selected-time > span,
 .mdp-timepicker-selected-ampm > span {
     outline: 0;
-    opacity: 0.6; 
-    
+    opacity: 0.6;
+
     &:not(.active) {
         cursor: pointer;
     }
-    
+
     &.active {
-        opacity: 1;    
+        opacity: 1;
     }
 }
