+++ src/components/mdpTimePicker/templates/clock.html
@@ -1,0 +1,17 @@
+<div class="mdp-clock">
+    <div class="mdp-clock-container">
+        <md-toolbar class="mdp-clock-center md-primary"></md-toolbar>
+
+        <md-toolbar ng-style="clock.getPointerStyle()" class="mdp-pointer md-primary">
+            <span class="mdp-clock-selected md-button md-raised md-primary"></span>
+        </md-toolbar>
+
+        <md-button
+        	ng-class="{'md-primary': clock.selected == step}"
+        	class="md-icon-button md-raised mdp-clock-deg{{ ::(clock.STEP_DEG * ($index + 1)) }}"
+        	ng-repeat="step in clock.steps"
+        	ng-click="clock.setTime(step)">
+        	{{ step }}
+        </md-button>
+    </div>
+</div>
