+++ src/core/style/animation.less
@@ -5,35 +5,34 @@
 .mdp-animate-next {
    opacity: 0;
    transform: translate3d(50%, 0, 1px);
-   
-   
+
     &-remove {
         transition: all 0.5s cubic-bezier(0.35, 0, 0.25, 1);
         opacity: 0;
-        transform: translate3d(50%, 0, 1px);    
+        transform: translate3d(50%, 0, 1px);
     }
 
     &-remove-active {
         opacity: 1;
         transform: translate3d(0, 0, 1px);
-    } 
+    }
 }
 
 .mdp-animate-prev {
    opacity: 0;
    transform: translate3d(-50%, 0, 1px);
-   
-   
+
+
     &-remove {
         transition: all 0.3s cubic-bezier(0.35, 0, 0.25, 1);
         opacity: 0;
-        transform: translate3d(-50%, 0, 1px);    
+        transform: translate3d(-50%, 0, 1px);
     }
 
     &-remove-active {
         opacity: 1;
         transform: translate3d(0, 0, 1px);
-    } 
+    }
 }
 
 @keyframes mdp-animation-bounce {
@@ -49,15 +48,15 @@
         transform: scale(1.0);
     }
 }
- 
+
 .mdp-animation-zoom {
     &.ng-enter {
         transition: all 0.3s cubic-bezier(0.35, 0, 0.25, 1);
-        animation-duration: 0.3s;        
+        animation-duration: 0.3s;
         animation-name: mdp-animation-bounce;
-    } 
-    
+    }
+
     &.ng-enter.ng-enter-active {
 
-    } 
+    }
 }
