+++ src/mdPickers.templates.js
@@ -1,0 +1,1 @@
+!function(e){try{e=angular.module("mdPickers")}catch(a){e=angular.module("mdPickers",[])}e.run(["$templateCache",function(e){e.put("mdpDatePicker/templates/mdp-date-picker-calendar.html",'<div class="mdp-calendar"><div layout="row" layout-align="space-between center"><md-button aria-label="previous month" class="md-icon-button" ng-click="calendar.prevMonth()"><md-icon md-svg-icon="mdp-chevron-left"></md-icon></md-button><div class="mdp-calendar-monthyear" ng-show="!calendar.animating">{{ calendar.date.format("MMMM YYYY") }}</div><md-button aria-label="next month" class="md-icon-button" ng-click="calendar.nextMonth()"><md-icon md-svg-icon="mdp-chevron-right"></md-icon></md-button></div><div layout="row" layout-align="space-around center" class="mdp-calendar-week-days" ng-show="!calendar.animating"><div layout="" layout-align="center center" ng-repeat="d in ::calendar.weekDays track by $index">{{ ::d }}</div></div><div layout="row" layout-align="start center" layout-wrap="" class="mdp-calendar-days" ng-class="{ \'mdp-animate-next\': calendar.animating }" ng-show="!calendar.animating" md-swipe-left="calendar.nextMonth()" md-swipe-right="calendar.prevMonth()"><div layout="" layout-align="center center" ng-repeat-start="day in calendar.daysInMonth track by $index" ng-class="{ \'mdp-day-placeholder\': !day }"><md-button class="md-icon-button md-raised" aria-label="Select day" ng-if="day" ng-class="{ \'md-accent\': calendar.date.date() == day.value }" ng-click="calendar.selectDate(day.value)" ng-disabled="!day.enabled">{{ ::day.value }}</md-button></div><div flex="100" ng-if="($index + 1) % 7 == 0" ng-repeat-end=""></div></div></div>')}])}(),function(e){try{e=angular.module("mdPickers")}catch(a){e=angular.module("mdPickers",[])}e.run(["$templateCache",function(e){e.put("mdpDatePicker/templates/mdp-date-picker.html",'<md-dialog class="mdp-datepicker" ng-class="{portrait: !$mdMedia(\'gt-xs\')}"><md-dialog-content layout="row" layout-wrap=""><div layout="column" layout-align="start center"><md-toolbar layout-align="start start" flex="noshrink" class="mdp-datepicker-date-wrapper md-hue-1 md-primary" layout="column"><span class="mdp-datepicker-year" ng-click="datepicker.showYear()" ng-class="{active: datepicker.selectingYear}">{{ datepicker.date.format(\'YYYY\') }}</span> <span class="mdp-datepicker-date" ng-click="datepicker.showCalendar()" ng-class="{active: !datepicker.selectingYear}">{{ datepicker.date.format(datepicker.displayFormat) }}</span></md-toolbar></div><div><div class="mdp-datepicker-select-year mdp-animation-zoom" layout="column" layout-align="center start" ng-if="datepicker.selectingYear"><md-virtual-repeat-container md-auto-shrink="" md-top-index="datepicker.yearTopIndex"><div flex="" md-virtual-repeat="item in ::datepicker.yearItems" md-on-demand="" class="repeated-year"><span class="md-button" ng-click="datepicker.selectYear(item)" md-ink-ripple="" ng-class="{ \'md-primary current\': item == year }">{{ item }}</span></div></md-virtual-repeat-container></div><mdp-calendar ng-if="!datepicker.selectingYear" class="mdp-animation-zoom" date="datepicker.date" min-date="datepicker.minDate" date-filter="datepicker.dateFilter" max-date="datepicker.maxDate"></mdp-calendar><md-dialog-actions layout="row"><span flex=""></span><md-button ng-click="datepicker.cancel()" aria-label="{{::datepicker.labels.cancel}}">{{::datepicker.labels.cancel}}</md-button><md-button ng-click="datepicker.confirm()" class="md-primary" aria-label="{{::datepicker.labels.ok}}">{{::datepicker.labels.ok}}</md-button></md-dialog-actions></div></md-dialog-content></md-dialog>')}])}(),function(e){try{e=angular.module("mdPickers")}catch(a){e=angular.module("mdPickers",[])}e.run(["$templateCache",function(e){e.put("mdpTimePicker/templates/clock.html",'<div class="mdp-clock"><div class="mdp-clock-container"><md-toolbar class="mdp-clock-center md-primary"></md-toolbar><md-toolbar ng-style="clock.getPointerStyle()" class="mdp-pointer md-primary"><span class="mdp-clock-selected md-button md-raised md-primary"></span></md-toolbar><md-button ng-class="{\'md-primary\': clock.selected == step}" class="md-icon-button md-raised mdp-clock-deg{{ ::(clock.STEP_DEG * ($index + 1)) }}" ng-repeat="step in clock.steps" ng-click="clock.setTime(step)">{{ step }}</md-button></div></div>')}])}(),function(e){try{e=angular.module("mdPickers")}catch(a){e=angular.module("mdPickers",[])}e.run(["$templateCache",function(e){e.put("mdpTimePicker/templates/mdp-time-picker.html",'<md-dialog class="mdp-timepicker" ng-class="{portrait: !$mdMedia(\'gt-xs\')}"><md-dialog-content layout="column" layout-gt-xs="row"><md-toolbar layout-gt-xs="column" layout-xs="row" layout-align="center center" flex="noshrink" class="mdp-timepicker-time md-hue-1 md-primary"><div class="mdp-timepicker-selected-time"><span ng-class="{ \'active\': timepicker.currentView == timepicker.VIEW_HOURS }" ng-click="timepicker.currentView = timepicker.VIEW_HOURS">{{ timepicker.time.format("h") }}</span>:<span ng-class="{ \'active\': timepicker.currentView == timepicker.VIEW_MINUTES }" ng-click="timepicker.currentView = timepicker.VIEW_MINUTES">{{ timepicker.time.format("mm") }}</span></div><div layout="column" class="mdp-timepicker-selected-ampm"><span ng-click="timepicker.setAM()" ng-class="{ \'active\': timepicker.time.hours() < 12 }">AM</span> <span ng-click="timepicker.setPM()" ng-class="{ \'active\': timepicker.time.hours() >= 12 }">PM</span></div></md-toolbar><div flex="noshrink"><div class="mdp-clock-switch-container" ng-switch="timepicker.currentView" layout="" layout-align="center center"><mdp-clock class="mdp-animation-zoom" auto-switch="timepicker.autoSwitch" time="timepicker.time" type="hours" ng-switch-when="1"></mdp-clock><mdp-clock class="mdp-animation-zoom" auto-switch="timepicker.autoSwitch" time="timepicker.time" type="minutes" ng-switch-when="2"></mdp-clock></div><md-dialog-actions layout="row"><span flex=""></span><md-button ng-click="timepicker.cancel()" aria-label="{{::timepicker.labels.cancel}}">{{::timepicker.labels.cancel}}</md-button><md-button ng-click="timepicker.confirm()" class="md-primary" aria-label="{{::timepicker.labels.ok}}">{{::timepicker.labels.ok}}</md-button></md-dialog-actions></div></md-dialog-content></md-dialog>')}])}();
